+++ src/main/java/cofh/core/util/helpers/vfx/VFXHelper.java
@@ -249,7 +249,6 @@ public final class VFXHelper {
     public static void renderTest(PoseStack stack, VertexConsumer consumer) {
 
         Vector4f center = new Vector4f(0, 0, 0, 1).mul(stack.last().pose());
-        // setNormal takes the Pose now; the bare Matrix3f overload is gone.
         PoseStack.Pose pose = stack.last();
         float xp = center.x() + 0.5F;
         float xn = center.x() - 0.5F;
@@ -261,10 +260,10 @@ public final class VFXHelper {
         int b = 255;
         int a = 255;
         int overlay = OverlayTexture.pack(0, false); //OverlayTexture.NO_OVERLAY
-        consumer.addVertex((float) (xp), (float) (yp), (float) (z)).setColor(r, g, b, a).setUv(0, 0).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
-        consumer.addVertex((float) (xn), (float) (yp), (float) (z)).setColor(r, g, b, a).setUv(0, 1).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
-        consumer.addVertex((float) (xn), (float) (yn), (float) (z)).setColor(r, g, b, a).setUv(1, 1).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
-        consumer.addVertex((float) (xp), (float) (yn), (float) (z)).setColor(r, g, b, a).setUv(1, 0).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
+        consumer.addVertex(xp, yp, z).setColor(r, g, b, a).setUv(0, 0).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
+        consumer.addVertex(xn, yp, z).setColor(r, g, b, a).setUv(0, 1).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
+        consumer.addVertex(xn, yn, z).setColor(r, g, b, a).setUv(1, 1).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
+        consumer.addVertex(xp, yn, z).setColor(r, g, b, a).setUv(1, 0).setOverlay(overlay).setLight(RenderHelper.FULL_BRIGHT).setNormal(pose, 0, 1, 0);
     }
 
     public static void renderTest(PoseStack stack, MultiBufferSource buffer) {
@@ -283,30 +282,30 @@ public final class VFXHelper {
         float xm = node.xMid();
         float ym = node.yMid();
 
-        builder.addVertex((float) (node.xp), (float) (node.yp), (float) (node.z)).setColor(255, 0, 0, 255);
-        builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 0, 0, 255);
-        builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(0, 0, 255, 255);
-        builder.addVertex((float) (node.xn), (float) (node.yn), (float) (node.z)).setColor(0, 0, 255, 255);
-        builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 255, 255, 255);
+        builder.addVertex(node.xp, node.yp, node.z).setColor(255, 0, 0, 255);
+        builder.addVertex(xm, ym, node.z).setColor(255, 0, 0, 255);
+        builder.addVertex(xm, ym, node.z).setColor(0, 0, 255, 255);
+        builder.addVertex(node.xn, node.yn, node.z).setColor(0, 0, 255, 255);
+        builder.addVertex(xm, ym, node.z).setColor(255, 255, 255, 255);
         for (int i = 1; i < nodes.length - 1; ++i) {
             node = nodes[i];
             xm = node.xMid();
             ym = node.yMid();
-            builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 255, 255, 255);
-            builder.addVertex((float) (node.xp), (float) (node.yp), (float) (node.z)).setColor(255, 0, 0, 255);
-            builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 0, 0, 255);
-            builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(0, 0, 255, 255);
-            builder.addVertex((float) (node.xn), (float) (node.yn), (float) (node.z)).setColor(0, 0, 255, 255);
-            builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 255, 255, 255);
+            builder.addVertex(xm, ym, node.z).setColor(255, 255, 255, 255);
+            builder.addVertex(node.xp, node.yp, node.z).setColor(255, 0, 0, 255);
+            builder.addVertex(xm, ym, node.z).setColor(255, 0, 0, 255);
+            builder.addVertex(xm, ym, node.z).setColor(0, 0, 255, 255);
+            builder.addVertex(node.xn, node.yn, node.z).setColor(0, 0, 255, 255);
+            builder.addVertex(xm, ym, node.z).setColor(255, 255, 255, 255);
         }
         node = nodes[nodes.length - 1];
         xm = node.xMid();
         ym = node.yMid();
-        builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 255, 255, 255);
-        builder.addVertex((float) (node.xp), (float) (node.yp), (float) (node.z)).setColor(255, 0, 0, 255);
-        builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(255, 0, 0, 255);
-        builder.addVertex((float) (xm), (float) (ym), (float) (node.z)).setColor(0, 0, 255, 255);
-        builder.addVertex((float) (node.xn), (float) (node.yn), (float) (node.z)).setColor(0, 0, 255, 255);
+        builder.addVertex(xm, ym, node.z).setColor(255, 255, 255, 255);
+        builder.addVertex(node.xp, node.yp, node.z).setColor(255, 0, 0, 255);
+        builder.addVertex(xm, ym, node.z).setColor(255, 0, 0, 255);
+        builder.addVertex(xm, ym, node.z).setColor(0, 0, 255, 255);
+        builder.addVertex(node.xn, node.yn, node.z).setColor(0, 0, 255, 255);
     }
 
     private static void renderSkeleton(VFXNode[] nodes) {
@@ -690,8 +689,8 @@ public final class VFXHelper {
 
         public VFXNode renderStart(Vector3f normal, VertexConsumer builder, int packedLight, Color col, float u0, float v0, float u1, float v1) {
 
-            builder.addVertex((float) (xp), (float) (yp), (float) (z)).setColor(col.r, col.g, col.b, col.a).setUv(u0, v0).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
-            builder.addVertex((float) (xn), (float) (yn), (float) (z)).setColor(col.r, col.g, col.b, col.a).setUv(u1, v0).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
+            builder.addVertex(xp, yp, z).setColor(col.r, col.g, col.b, col.a).setUv(u0, v0).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
+            builder.addVertex(xn, yn, z).setColor(col.r, col.g, col.b, col.a).setUv(u1, v0).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
             return this;
         }
 
@@ -702,8 +701,8 @@ public final class VFXHelper {
 
         public VFXNode renderEnd(Vector3f normal, VertexConsumer builder, int packedLight, Color col, float u0, float v0, float u1, float v1) {
 
-            builder.addVertex((float) (xn), (float) (yn), (float) (z)).setColor(col.r, col.g, col.b, col.a).setUv(u1, v1).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
-            builder.addVertex((float) (xp), (float) (yp), (float) (z)).setColor(col.r, col.g, col.b, col.a).setUv(u0, v1).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
+            builder.addVertex(xn, yn, z).setColor(col.r, col.g, col.b, col.a).setUv(u1, v1).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
+            builder.addVertex(xp, yp, z).setColor(col.r, col.g, col.b, col.a).setUv(u0, v1).setOverlay(OverlayTexture.NO_OVERLAY).setLight(packedLight).setNormal(normal.x, normal.y, normal.z);
             return this;
         }
